Add selective disclosure for JWT credentials
- Add basic SD example, for FE-sized attribs

- Attributes that are larger than a field element can be hashed to fit,
then disclosed by revealing the preimage. Completes the core work for
selective disclosure.  Added support for basic proof specifications
encoding a description of attribute data to disclose.

- Callers do not have to specify whether a revaled attribute is hashed or
not, we can figure that out from the config.json file.

- The Groth16 params were being saved twice, once on their own and once in
the prover params, as they can be large, we only save them once in the
prover params.

- Add "prepare" option to command line tool, as a synonym for "prove"

Signed-off-by: Greg Zaverucha <gregz@microsoft.com>