Simplify proof spec in public API
Callers do not have to specify whether a revaled attribute is hashed or
not, we can figure that out from the config.json file.

The Groth16 params were being saved twice, once on their own and once in
the prover params, as they can be large, we only save them once in the
prover params.

Signed-off-by: Greg Zaverucha <gregz@microsoft.com>