Add sub-prover for device bound credentials
The new sub-prover takes a commitment to the public key from the main
show proof, together with an ECDSA signature produced by the device and
and creates a proof of posession of the signture.  This feature is
necessary to support device-bound credentials, such as mDLs.

Currently the sub-prover can only be tested in isolation, it mocks the
commitments it expects from the main show proof code.   Next steps will
be to integrate it with the rest of Crescent.

Signed-off-by: Greg Zaverucha <gregz@microsoft.com>